create class for convert_external_network_injections_cim16
